+++ web/js/common.js
@@ -549,9 +549,17 @@ function setEndTime($fromHour, $fromMinute, $toHour, $toMinute, $nbMinutes) {
  */
 function timesAreValid(startHour, endHour, startMinute, endMinute) {
     if (startHour.length && endHour.length) {
-        if (!$.isNumeric(startHour.val()) || !$.isNumeric(endHour.val()) ||
-            !$.isNumeric(startMinute.val()) || !$.isNumeric(endMinute.val())) {
+        if ($.isNumeric(startHour.val()) && $.isNumeric(endHour.val()) &&
+            $.isNumeric(startMinute.val()) && $.isNumeric(endMinute.val())) {
+            var startHourVal = parseInt(startHour.val());
+            var endHourVal = parseInt(endHour.val());
 
+            //if (startHourVal > endHourVal && endHourVal != '0') {//To test: Should be the solution to avoid error when booking duration is less than 1 hour
+            if (startHourVal >= endHourVal && endHourVal != '0') {
+                $("#time-error").show();
+                return false;
+            }
+        } else {
             return false;
         }
     }